Fix fletcher16 checksum computation.
In the process of adding the more elaborate tests in this commit, I
found that fletcher16 was not performing well.  Out of the 2^16 random
single bit-flips I tried, 25 had the same checksum.  And one particular
checksum value had ~128/2^16 preimages, i.e. 128 of the 2^16 inputs we
tried all hashed to the same value.

After a bit of investigation, I realized that this was all due to a
subtle C typecasting bug:

    *sum1 = ((*sum1) + data[index]) % 255;

The issue is that `data` is a `char`, which may be signed or unsigned
depending on the platform.  If it's signed, then this is bad!

This commit fixes the problem, adds tests that catch the problem
statistically and explicitly, and also simplifies the code and API.