Revert "Removed time checks from PID library."
This reverts commit be574eaddb12501e9bfff7f74b8f7210dbe554af.

The time checks were removed from PID because we were only calling PID
from the precisely-timed control loop.  We tried to update the PID every
(say) 10ms, and we were confident we *would* call every 10ms, plus or
minus a few microseconds.  But if we happened to call with a delta of
9.99ms, the PID code would skip our sample!  That's bad.

However, for flow correction, we need our PID library to be able to
handle the case when the time interval is not constant.  In this case,
we *do* want time checks.

So in this change we add time checks back.  The next commit makes time
checking optional, so the PID library can work for both cases.

GPC: pid-time